(edit) extract the Save and Go Back button into its own reusable component, DatasetSaveAndGoBack
